requirements.txt
cv2
numpy
torch
pathlib
pillow
typing
ultralytics
threading